requirements.txt
# FastAPI Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
python-multipart==0.0.20

# Image Processing
Pillow==10.4.0

# OCR Engine
rapidocr-onnxruntime>=1.3.0  # PaddleOCR (한국어 텍스트 인식)

# LLM Text Correction (Hugging Face API)
requests>=2.31.0             # Lightweight HTTP client for HF API
